# src/backend/core/api/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Deploy Center API",
        "version": "1.0.0 (v1.0)",
        "description": "This is the Deploy Center API schema."
    },
    "paths": {
...
        "/api/v1.0/operators/{operator_id}/organizations/{id}/operator-role/": {
            "patch": {
                "operationId": "operators_organizations_operator_role_partial_update",
                "description": "Update the OperatorOrganizationRole settings for this org+operator.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "description": "primary key for the record as UUID"
                        },
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "operator_id",
                        "schema": {
                            "type": "string",
                            "pattern": "^[0-9a-z-]*$"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "operators"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Organization"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1.0/operators/{operator_id}/organizations/{id}/proconnect-domains/": {
            "patch": {
                "operationId": "operators_organizations_proconnect_domains_partial_update",
                "description": "Update the organization's ProConnect domain buckets.\n\nBody may contain \"manual\", \"requested\" and/or \"discarded\" lists (partial):\n- \"requested\": any operator member — domains requested for a superuser to\n  validate (i.e. move to \"manual\").\n- \"manual\" / \"discarded\": superuser only.\n\nThe system-managed buckets (\"dpnt\", \"candidates\") are preserved.\nEach change is recorded as an admin LogEntry on the organization.",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "description": "primary key for the record as UUID"
                        },
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "operator_id",
                        "schema": {
                            "type": "string",
                            "pattern": "^[0-9a-z-]*$"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "operators"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedProConnectDomainsUpdateRequest"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProConnectDomains"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Entitlement": {
                "type": "object",
                "description": "Serialize entitlements.",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true,
                        "description": "primary key for the record as UUID"
                    },
                    "type": {
                        "type": "string",
                        "readOnly": true,
                        "description": "Type of entitlement"
                    },
                    "config": {
                        "nullable": true,
                        "title": "Configuration",
                        "description": "Base configuration data for metrics scraping and service operation"
                    },
                    "account_type": {
                        "type": "string",
                        "description": "Type of account",
                        "maxLength": 50
                    },
                    "account": {
                        "type": "string",
                        "format": "uuid",
                        "description": "Account this entitlement is associated with",
                        "nullable": true
                    }
                },
                "required": [
                    "id",
                    "type"
                ]
            },
...
            "Operator": {
                "type": "object",
                "description": "Serialize operators.",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true,
                        "description": "primary key for the record as UUID"
                    },
                    "name": {
                        "type": "string",
                        "readOnly": true,
                        "title": "Nom",
                        "description": "Name of the operator organization"
                    },
                    "siret": {
                        "type": "string",
                        "readOnly": true,
                        "description": "SIRET number of the operator organization"
                    },
                    "url": {
                        "type": "string",
                        "format": "uri",
                        "readOnly": true,
                        "nullable": true,
                        "title": "Homepage",
                        "description": "Homepage URL of the operator"
                    },
                    "is_active": {
                        "type": "boolean",
                        "readOnly": true,
                        "title": "Actif",
                        "description": "Whether this operator is currently active"
                    },
                    "user_role": {
                        "type": "string",
                        "readOnly": true
                    },
                    "config": {
                        "type": "string",
                        "readOnly": true
                    }
                },
                "required": [
                    "config",
                    "id",
                    "is_active",
                    "name",
                    "siret",
                    "url",
                    "user_role"
                ]
            },
            "Organization": {
                "type": "object",
                "description": "Serialize organizations.",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true,
                        "description": "primary key for the record as UUID"
                    },
                    "name": {
                        "type": "string",
                        "readOnly": true,
                        "title": "Nom",
                        "description": "Official name of the collectivité"
                    },
                    "type": {
                        "type": "string",
                        "readOnly": true,
                        "description": "Type of collectivité (commune, epci, departement, region)"
                    },
                    "siret": {
                        "type": "string",
                        "readOnly": true,
                        "nullable": true,
                        "description": "SIRET code of the main establishment"
                    },
                    "siren": {
                        "type": "string",
                        "readOnly": true,
                        "nullable": true,
                        "description": "SIREN code of the organization"
                    },
                    "code_postal": {
                        "type": "string",
                        "readOnly": true,
                        "nullable": true,
                        "title": "Postal code",
                        "description": "Postal code of the collectivité"
                    },
                    "code_insee": {
                        "type": "string",
                        "readOnly": true,
                        "nullable": true,
                        "title": "INSEE code",
                        "description": "INSEE administrative code"
                    },
                    "population": {
                        "type": "integer",
                        "readOnly": true,
                        "nullable": true,
                        "description": "Population count from latest INSEE data"
                    },
                    "epci_libelle": {
                        "type": "string",
                        "readOnly": true,
                        "nullable": true,
                        "title": "EPCI name",
                        "description": "Name of the EPCI the commune belongs to"
                    },
                    "epci_siren": {
                        "type": "string",
                        "readOnly": true,
                        "nullable": true,
                        "description": "SIREN code of the EPCI"
                    },
                    "epci_population": {
                        "type": "integer",
                        "readOnly": true,
                        "nullable": true,
                        "description": "Population of the EPCI"
                    },
                    "departement_code_insee": {
                        "type": "string",
                        "readOnly": true,
                        "nullable": true,
                        "title": "Department code INSEE",
                        "description": "INSEE department code"
                    },
                    "region_code_insee": {
                        "type": "string",
                        "readOnly": true,
                        "nullable": true,
                        "description": "INSEE region code"
                    },
                    "adresse_messagerie": {
                        "type": "string",
                        "format": "email",
                        "readOnly": true,
                        "nullable": true,
                        "title": "Adresse électronique",
                        "description": "Official email address from Service-Public.fr"
                    },
                    "site_internet": {
                        "type": "string",
                        "format": "uri",
                        "readOnly": true,
                        "nullable": true,
                        "title": "Website",
                        "description": "Official website URL from Service-Public.fr"
                    },
                    "telephone": {
                        "type": "string",
                        "readOnly": true,
                        "nullable": true,
                        "title": "Phone",
                        "description": "Official phone number from Service-Public.fr"
                    },
                    "rpnt": {
                        "readOnly": true,
                        "nullable": true,
                        "title": "RPNT criteria",
                        "description": "List of valid RPNT criteria and meta-criteria"
                    },
                    "service_public_url": {
                        "type": "string",
                        "format": "uri",
                        "readOnly": true,
                        "nullable": true,
                        "description": "URL of the associated page on Service-Public.fr"
                    },
                    "proconnect_domains": {
                        "type": "object",
                        "description": "The org's ProConnect domains by provenance (dpnt/candidates/manual) and status (requested/discarded).",
                        "properties": {
                            "requested": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "manual": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "dpnt": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "candidates": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            },
                            "discarded": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        },
                        "readOnly": true
                    },
                    "proconnect_routable": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "description": "The domains the org may route to a ProConnect provider — the buckets with the discard rules applied. The UI offers exactly these.",
                        "readOnly": true
                    },
                    "proconnect_prevalidated": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "description": "Per idp with a known deployed allowlist, the org domains already in it. Null when no allowlist is known (pre-validation unknown).",
                        "readOnly": true
                    },
                    "service_subscriptions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ServiceSubscriptionWithService"
                        },
                        "readOnly": true
                    }
                },
                "required": [
                    "adresse_messagerie",
                    "code_insee",
                    "code_postal",
                    "departement_code_insee",
                    "epci_libelle",
                    "epci_population",
                    "epci_siren",
                    "id",
                    "name",
                    "population",
                    "proconnect_domains",
                    "proconnect_prevalidated",
                    "proconnect_routable",
                    "region_code_insee",
                    "rpnt",
                    "service_public_url",
                    "service_subscriptions",
                    "siren",
                    "siret",
                    "site_internet",
                    "telephone",
                    "type"
                ]
            },
...
            "PatchedProConnectDomainsUpdateRequest": {
                "type": "object",
                "description": "Body of the ``proconnect-domains`` PATCH action.\n\nDocumentation/schema only — the action validates and normalizes the buckets\nitself (each key is optional; an omitted bucket is left untouched).",
                "properties": {
                    "manual": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "minLength": 1
                        }
                    },
                    "requested": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "minLength": 1
                        }
                    },
                    "discarded": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "minLength": 1
                        }
                    }
                }
            },
...
            "ProConnectDomains": {
                "type": "object",
                "description": "The normalized domain buckets returned by the ``proconnect-domains`` action.\n\nDocumentation/schema only — the action returns the dict built by the service\nlayer, not a serialized model.",
                "properties": {
                    "requested": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "manual": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "dpnt": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "candidates": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "discarded": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "candidates",
                    "discarded",
                    "dpnt",
                    "manual",
                    "requested"
                ]
            },
...
            "ServiceSubscriptionWithService": {
                "type": "object",
                "description": "Serialize service subscriptions with the service.",
                "properties": {
                    "metadata": {
                        "description": "Additional metadata for this subscription"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "title": "Created on",
                        "description": "date and time at which a record was created"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "title": "Updated on",
                        "description": "date and time at which a record was last updated"
                    },
                    "is_active": {
                        "type": "boolean",
                        "readOnly": true,
                        "title": "Actif",
                        "description": "Whether this subscription is currently active"
                    },
                    "entitlements": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Entitlement"
                        },
                        "readOnly": true
                    },
                    "service": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Service"
                            }
                        ],
                        "readOnly": true
                    },
                    "operator": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/Operator"
                            }
                        ],
                        "readOnly": true
                    }
                },
                "required": [
                    "created_at",
                    "entitlements",
                    "is_active",
                    "operator",
                    "service",
                    "updated_at"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            }
        }
    }
}